refactor(fe): migrate barrel imports to @/modules/*, delete 8 barrels, clean empty dirs
- Migrate 7 store barrels (@/stores/* → @/modules/*/stores/*): 25 files updated
- Migrate dweller model barrel (@/models/dweller → @/modules/dwellers/models/dweller)
- Delete 8 barrel files (7 stores + 1 model) and 8 empty directories
- Fix 1 remaining legacy import (@/types/api.generated → @/core/types/api.generated)
- Verify zero legacy import paths remain

Tests: 857 pass, 1 skipped, 51 files green
