--- frontend/src/models/dweller.ts
@@ -1,10 +1,0 @@
-// Re-export from the dwellers module for backward compatibility
-// This file will be removed in a future cleanup when all imports are updated
-
-export type {
-  Dweller,
-  DwellerShort,
-  DwellerFull,
-  VisualAttributes,
-  Special,
-} from '@/modules/dwellers/models/dweller'
