+++ frontend/src/modules/chat/components/DwellerChatPage.vue
@@ -7,7 +7,7 @@ import { useVaultStore } from '@/modules/vault/stores/vault'
 import { UButton, UCard } from '@/core/components/ui'
 import { Icon } from '@iconify/vue'
 import DwellerChat from './DwellerChat.vue'
-import type { Dweller } from '@/models/dweller'
+import type { Dweller } from '@/modules/dwellers/models/dweller'
 import { normalizeImageUrl } from '@/utils/image'
 
 const route = useRoute()
