+++ frontend/src/modules/dwellers/views/DwellerDetailView.vue
@@ -4,7 +4,7 @@ import { useRoute, useRouter } from 'vue-router'
 import { useDwellerStore } from '../stores/dweller'
 import { useAuthStore } from '@/modules/auth/stores/auth'
 import { useVaultStore } from '@/modules/vault/stores/vault'
-import { useExplorationStore } from '@/stores/exploration'
+import { useExplorationStore } from '@/modules/exploration/stores/exploration'
 import { Icon } from '@iconify/vue'
 import SidePanel from '@/core/components/common/SidePanel.vue'
 import DwellerCard from '../components/cards/DwellerCard.vue'
