+++ frontend/src/modules/exploration/components/EventTimeline.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { computed } from 'vue'
 import { Icon } from '@iconify/vue'
-import type { Exploration, ExplorationEvent } from '@/stores/exploration'
+import type { Exploration, ExplorationEvent } from '@/modules/exploration/stores/exploration'
 import type { Dweller } from '@/modules/dwellers/stores/dweller'
 
 interface Props {
