+++ frontend/src/modules/progression/components/training/TrainingQueuePanel.vue
@@ -2,7 +2,7 @@
 import { computed, onMounted, onUnmounted, ref } from 'vue'
 import { Icon } from '@iconify/vue'
 import TrainingProgressCard from './TrainingProgressCard.vue'
-import { useTrainingStore } from '@/stores/training'
+import { useTrainingStore } from '@/modules/progression/stores/training'
 import { useAuthStore } from '@/modules/auth/stores/auth'
 import { useVaultStore } from '@/modules/vault/stores/vault'
