+++ frontend/src/modules/vault/views/HappinessView.vue
@@ -3,7 +3,7 @@ import { computed, onMounted, ref } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import { useVaultStore } from '../stores/vault'
 import { useDwellerStore } from '@/modules/dwellers/stores/dweller'
-import { useIncidentStore } from '@/stores/incident'
+import { useIncidentStore } from '@/modules/combat/stores/incident'
 import { useAuthStore } from '@/modules/auth/stores/auth'
 import { useSidePanel } from '@/core/composables/useSidePanel'
 import SidePanel from '@/core/components/common/SidePanel.vue'
