--- frontend/src/stores/equipment.ts
@@ -1,2 +1,0 @@
-// Re-export from combat module for backward compatibility
-export { useEquipmentStore } from '@/modules/combat/stores/equipment'
