--- frontend/src/stores/exploration.ts
@@ -1,9 +1,0 @@
-// Re-export from exploration module for backward compatibility
-export {
-  useExplorationStore,
-  type ExplorationEvent,
-  type LootItem,
-  type Exploration,
-  type ExplorationProgress,
-  type RewardsSummary,
-} from '@/modules/exploration/stores/exploration'
