--- frontend/src/stores/incident.ts
@@ -1,2 +1,0 @@
-// Re-export from combat module for backward compatibility
-export { useIncidentStore } from '@/modules/combat/stores/incident'
