--- frontend/src/stores/objectives.ts
@@ -1,2 +1,0 @@
-// Re-export from progression module for backward compatibility
-export { useObjectivesStore } from '@/modules/progression/stores/objectives'
