--- frontend/src/stores/quest.ts
@@ -1,2 +1,0 @@
-// Re-export from progression module for backward compatibility
-export { useQuestStore } from '@/modules/progression/stores/quest'
