--- frontend/src/stores/room.ts
@@ -1,2 +1,0 @@
-// Re-export from rooms module for backward compatibility
-export { useRoomStore } from '@/modules/rooms/stores/room'
