--- frontend/src/stores/training.ts
@@ -1,2 +1,0 @@
-// Re-export from progression module for backward compatibility
-export { useTrainingStore } from '@/modules/progression/stores/training'
