+++ frontend/tests/unit/stores/objectives.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest'
 import { setActivePinia, createPinia } from 'pinia'
-import { useObjectivesStore } from '@/stores/objectives'
+import { useObjectivesStore } from '@/modules/progression/stores/objectives'
 import axios from '@/core/plugins/axios'
 
 vi.mock('@/core/plugins/axios')
