+++ frontend/tests/unit/stores/quest.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { setActivePinia, createPinia } from 'pinia'
-import { useQuestStore } from '@/stores/quest'
+import { useQuestStore } from '@/modules/progression/stores/quest'
 import axios from '@/core/plugins/axios'
 
 vi.mock('@/core/plugins/axios')
