+++ frontend/tests/unit/views/DwellersView.test.ts
@@ -6,7 +6,7 @@ import DwellersView from '@/modules/dwellers/views/DwellersView.vue'
 import { useAuthStore } from '@/modules/auth/stores/auth'
 import { useDwellerStore } from '@/modules/dwellers/stores/dweller'
 import { useVaultStore } from '@/modules/vault/stores/vault'
-import { useRoomStore } from '@/stores/room'
+import { useRoomStore } from '@/modules/rooms/stores/room'
 import axios from '@/core/plugins/axios'
 
 vi.mock('@/core/plugins/axios')
